ai_module/pyproject.toml
[tool.poetry]
name = "photosearch"
version = "0.1.0"
description = "Search photos with prompt"
authors = ["Milutin-S <studen.m98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.110.0"}
transformers = {extras = ["torch"], version = "^4.38.1"}
pillow = "^10.2.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
